# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "This is the API documentation for my Laravel application.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/receiver_customers": {
            "get": {
                "tags": [
                    "Receiver Customer"
                ],
                "summary": "Retrieve a list of all receiver customers",
                "operationId": "039ed3184b2199a8ff11b26caffba8c6",
                "responses": {
                    "200": {
                        "description": "Receiver customer list retrieved successfully"
                    },
                    "404": {
                        "description": "Receiver customer list not found"
                    }
                }
            }
        },
        "/api/v1/receiver_customer_register": {
            "post": {
                "tags": [
                    "Receiver Customer"
                ],
                "summary": "Store a new receiver customer",
                "operationId": "b6c4d4bb44337b969099bb4744449adc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Receiver customer created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/receiver_customer_update/{id}": {
            "put": {
                "tags": [
                    "Receiver Customer"
                ],
                "summary": "Update a receiver customer",
                "operationId": "89eb801d3c0d07a2e379e040b0d93a64",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the receiver customer to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Receiver customer updated successfully"
                    },
                    "404": {
                        "description": "Receiver customer not found"
                    }
                }
            }
        },
        "/api/v1/receiver_customer_destroy/{id}": {
            "delete": {
                "tags": [
                    "Receiver Customer"
                ],
                "summary": "Delete a receiver customer",
                "operationId": "50c5208c813026ca91c48ef5c26465e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the receiver customer to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receiver customer deleted successfully"
                    },
                    "404": {
                        "description": "Receiver customer not found"
                    }
                }
            }
        },
        "/api/v1/receiver_customer_show/{id}": {
            "get": {
                "tags": [
                    "Receiver Customer"
                ],
                "summary": "Retrieve a specific receiver customer",
                "operationId": "aecb0d34ca7705d7e9978a6703308263",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the receiver customer to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receiver customer retrieved successfully"
                    },
                    "404": {
                        "description": "Receiver customer not found"
                    }
                }
            }
        },
...
    },
...
}